Execute last handshake steps out of Receiver thread
